+++ src/tada.ads
@@ -5,13 +5,12 @@ with Ada.Containers.Indefinite_Holders;
 package Tada is
    use Ada;
 
-   Version : constant String := "0.7.0";
+   Version : constant String := "0.8.0";
 
    Package_Index_Url : constant String :=
      "https://raw.githubusercontent.com/tadapm/tada-packages/refs/heads/main/index";
 
-   type Package_Kind is (Exe,
-                         Lib);
+   type Package_Kind is (Exe, Lib);
 
    package String_Holders is new Ada.Containers.Indefinite_Holders (String);